removing explicit conversion to float since autocast should already take care of this

+++ torch_harmonics/_disco_convolution.py
@@ -85,7 +85,7 @@ class _DiscoS2TransposeContractionCuda(torch.autograd.Function):
     def backward(ctx, grad_output):
         roff_idx, ker_idx, row_idx, col_idx, vals = ctx.saved_tensors
         inp_type = grad_output.dtype
-        grad_input = disco_cuda_extension.forward(grad_output.to(torch.float32).contiguous(), roff_idx, ker_idx, row_idx, col_idx, vals,
+        grad_input = disco_cuda_extension.forward(grad_output.contiguous(), roff_idx, ker_idx, row_idx, col_idx, vals,
                                         ctx.kernel_size, ctx.nlat_in, ctx.nlon_in)
         grad_input = grad_input.to(dtype=inp_type)
